Import time improvements
