+++ dimos/protocol/pubsub/impl/lcmpubsub.py
@@ -22,7 +22,6 @@ from typing import Any
 
 from dimos.msgs.protocol import DimosMsg
 from dimos.protocol.pubsub.encoders import (
-    JpegEncoderMixin,
     LCMEncoderMixin,
     PickleEncoderMixin,
 )
@@ -149,16 +148,9 @@ class PickleLCM(  # type: ignore[misc]
 ): ...
 
 
-class JpegLCM(  # type: ignore[misc]
-    JpegEncoderMixin,  # type: ignore[type-arg]
-    LCMPubSubBase,
-): ...
-
-
 __all__ = [
     "LCM",
     "Glob",
-    "JpegLCM",
     "LCMEncoderMixin",
     "LCMPubSubBase",
     "PickleLCM",
